+++ tests/it/cluster/cluster_test.go
@@ -30,12 +30,17 @@ func TestMain(m *testing.M) {
 func TestAppendLogs(t *testing.T) {
 	const numAppend = 100
 
+	logger, err := zap.NewDevelopment()
+	assert.NoError(t, err)
+	defer logger.Sync()
+
 	clus := it.NewVarlogCluster(t,
 		it.WithReplicationFactor(2),
 		it.WithNumberOfStorageNodes(2),
 		it.WithNumberOfLogStreams(10),
 		it.WithNumberOfClients(10),
 		it.WithNumberOfTopics(1),
+		it.WithLogger(logger),
 	)
 	defer clus.Close(t)
 
@@ -57,6 +62,7 @@ func TestAppendLogs(t *testing.T) {
 				}
 				max = res.Metadata[0].GLSN
 			}
+			t.Logf("[%d] Append completed", idx)
 
 			muMaxGLSN.Lock()
 			defer muMaxGLSN.Unlock()